Dev/master

+++ praca_inz/praca_inz-web/src/main/web/src/app/layout/admin/components/administration/company/company-list/company-list.component.ts
@@ -1,30 +1,36 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 
 import {TranslateService} from "@ngx-translate/core";
 import {CompanyService} from "../company.service";
 import {StructureListElement} from "../../../../../../models/structure-elements";
+import {MessageService} from "../../../../../../shared/services/message.service";
+import {Configuration} from "../../../../../../app.constants";
 
 
 @Component({
-  selector: 'app-company-list',
-  templateUrl: './company-list.component.html',
-  styleUrls: ['./company-list.component.scss']
+    selector: 'app-company-list',
+    templateUrl: './company-list.component.html',
+    styleUrls: ['./company-list.component.scss']
 })
 export class CompanyListComponent implements OnInit {
 
     public deletedFilter = false;
     companies: StructureListElement[];
 
-    constructor(private companyService : CompanyService,
-                private translate:TranslateService,
-    ) {}
+    constructor(
+        private companyService: CompanyService,
+        private translate: TranslateService,
+        private messageService:MessageService,
+        private configuration:Configuration
+        ) {
+    }
 
     ngOnInit() {
         this.filterCompanies(null);
     }
 
-    getAddress(company:StructureListElement): string {
-        if (company.flatNumber== null || company.flatNumber === "0") {
+    getAddress(company: StructureListElement): string {
+        if (company.flatNumber == null || company.flatNumber === "0") {
             return (company.street + ' ' + company.buildingNumber);
         } else {
             return (company.street + ' ' + company.buildingNumber + ' / ' + company.flatNumber);
@@ -43,7 +49,7 @@ export class CompanyListComponent implements OnInit {
 
             searchText = searchText.toLowerCase();
             this.companies = copmanies.filter(it => {
-                const range = it.name + ' ' + it.street+' '+it.city+' '+it.description+' '+it.flatNumber+' '+it.buildingNumber;
+                const range = it.name + ' ' + it.street + ' ' + it.city + ' ' + it.description + ' ' + it.flatNumber + ' ' + it.buildingNumber;
                 const ok = range.toLowerCase().includes(searchText);
                 return ok;
             });
@@ -51,8 +57,28 @@ export class CompanyListComponent implements OnInit {
     }
 
     delete(structure: StructureListElement) {
-        this.companyService.deleteCompany(String(structure.id)).subscribe(resp => {
-            this.filterCompanies(null)
-        });
+        var entity: string;
+        var message: string;
+        var yes: string;
+        var no: string;
+
+        this.translate.get('comapny.delete').subscribe(x => entity = x);
+        this.translate.get('confirm.delete').subscribe(x => message = x);
+        this.translate.get('yes').subscribe(x => yes = x);
+        this.translate.get('no').subscribe(x => no = x);
+
+
+        this.messageService
+            .confirm(entity, message, yes, no)
+            .subscribe(confirmed => {
+                if (confirmed) {
+                    this.companyService.deleteCompany(String(structure.id)).subscribe(resp => {
+                        this.filterCompanies(null)
+                        this.translate.get('success.company.delete').subscribe(x=>{
+                            this.messageService.success(x)
+                        })
+                    });
+                }
+            });
     }
 }
