+++ praca_inz/praca_inz-web/src/main/web/src/app/layout/admin/components/structure-management/department/department-list/department-list.component.ts
@@ -5,6 +5,8 @@ import {StructureListElement} from "../../../../../../models/structure-elements"
 import {UserRoles} from "../../../../../../models/user-roles";
 import {UserService} from "../../../administration/user-management/user.service";
 import {LoggedUser} from "../../../../../../models/logged-user";
+import {MessageService} from "../../../../../../shared/services/message.service";
+import {Configuration} from "../../../../../../app.constants";
 
 
 @Component({
@@ -13,13 +15,20 @@ import {LoggedUser} from "../../../../../../models/logged-user";
     styleUrls: ['./department-list.component.scss']
 })
 export class DepartmentListComponent implements OnInit {
-    
+
     departments: StructureListElement[];
 
     roles: UserRoles;
     currentUser: LoggedUser;
 
-    constructor(private departmentService: DepartmentService, private userService: UserService, private translate: TranslateService) {
+    constructor(
+        private departmentService: DepartmentService,
+        private userService: UserService,
+        private translate: TranslateService,
+        private configuration: Configuration,
+        private messageService: MessageService
+
+    ) {
     }
 
     ngOnInit() {
@@ -30,7 +39,8 @@ export class DepartmentListComponent implements OnInit {
     }
 
 
-    getAddress(department: StructureListElement): string {
+    getAddress(department: StructureListElement):
+        string {
         if (department.flatNumber == null || department.flatNumber === "0") {
             return (department.street + ' ' + department.buildingNumber);
         } else {
@@ -39,11 +49,11 @@ export class DepartmentListComponent implements OnInit {
     }
 
     getLoggedUser() {
-        this.userService.getLoggedUser().subscribe(x => this.currentUser = x);
+        this.userService.getLoggedUser().subscribe(x => {this.currentUser = x});
     }
 
     getLoggedUserRoles() {
-        this.userService.getLoggedUserRoles().subscribe(x => this.roles = x);
+        this.userService.getLoggedUserRoles().subscribe(x => {this.roles = x});
     }
 
     filterDepartments(searchText: string) {
@@ -58,7 +68,7 @@ export class DepartmentListComponent implements OnInit {
 
             searchText = searchText.toLowerCase();
             this.departments = departments.filter(it => {
-                const range = it.name + ' '+ it.street+' '+it.city+' '+it.description+' '+it.flatNumber+' '+it.buildingNumber+' '+it.companyName;
+                const range = it.name + ' ' + it.street + ' ' + it.city + ' ' + it.description + ' ' + it.flatNumber + ' ' + it.buildingNumber + ' ' + it.companyName;
                 const ok = range.toLowerCase().includes(searchText);
                 return ok;
             });
@@ -66,9 +76,28 @@ export class DepartmentListComponent implements OnInit {
     }
 
     delete(structure: StructureListElement) {
-        this.departmentService.deleteDepartament(String(structure.id)).subscribe(resp => {
-            this.filterDepartments(null);
-        });
-    }
+        var entity: string;
+        var message: string;
+        var yes: string;
+        var no: string;
 
+        this.translate.get('department.delete').subscribe(x => entity = x);
+        this.translate.get('confirm.delete').subscribe(x => message = x);
+        this.translate.get('yes').subscribe(x => yes = x);
+        this.translate.get('no').subscribe(x => no = x);
+
+
+        this.messageService
+            .confirm(entity, message, yes, no)
+            .subscribe(confirmed => {
+                if (confirmed) {
+                    this.departmentService.deleteDepartament(String(structure.id)).subscribe(resp => {
+                        this.filterDepartments(null);
+                        this.translate.get('success.department.delete').subscribe(x=>{
+                            this.messageService.success(x)
+                        })
+                    });
+                }
+            });
+    }
 }
