+++ praca_inz/praca_inz-web/src/main/web/src/app/layout/admin/components/structure-management/device/device-model-list/device-model-list.component.ts
@@ -1,11 +1,10 @@
 import {Component, OnInit} from '@angular/core';
 import {DeviceService} from "../../../../../device-management/device.service";
 import {TranslateService} from "@ngx-translate/core";
-import {DeviceListElement, DeviceModelListElement} from "../../../../../../models/device-elements";
-import {StructureListElement} from "../../../../../../models/structure-elements";
-import {UserRoles} from "../../../../../../models/user-roles";
-import {LoggedUser} from "../../../../../../models/logged-user";
+import {DeviceModelListElement} from "../../../../../../models/device-elements";
 import {UserService} from "../../../administration/user-management/user.service";
+import {Configuration} from "../../../../../../app.constants";
+import {MessageService} from "../../../../../../shared/services/message.service";
 
 @Component({
     selector: 'app-device-model-list',
@@ -19,7 +18,9 @@ export class DeviceModelListComponent implements OnInit {
     constructor(
         private deviceService: DeviceService,
         private translate: TranslateService,
-        private userService: UserService
+        private userService: UserService,
+        private configuration: Configuration,
+        private messageService: MessageService
     ) {
     }
 
@@ -29,9 +30,9 @@ export class DeviceModelListComponent implements OnInit {
 
 
     getDevicesModels() {
-            this.deviceService.getAllDevicesModels().subscribe(deviceListElement => {
-                this.deviceModels = deviceListElement
-            });
+        this.deviceService.getAllDevicesModels().subscribe(deviceListElement => {
+            this.deviceModels = deviceListElement
+        });
     }
 
     filterDeviceModels(searchText: string) {
@@ -46,19 +47,42 @@ export class DeviceModelListComponent implements OnInit {
 
             searchText = searchText.toLowerCase();
             this.deviceModels = deviceModels.filter(it => {
-                const range = it.name + ' ' + it.manufacture + ' ' + it.type;
+                const range = it.name + ' ' + it.manufacture + ' ' + it.deviceTypeName;
                 const ok = range.toLowerCase().includes(searchText);
                 return ok;
             });
         });
     }
 
-    delete(deviceModel
-               :
-               DeviceModelListElement
-    ) {
-        this.deviceService.deleteDeviceModel(String(deviceModel.id)).subscribe(resp => {
-            this.getDevicesModels()
-        });
+    delete(deviceModel: DeviceModelListElement) {
+        var entity: string;
+        var message: string;
+        var yes: string;
+        var no: string;
+
+        this.translate.get('device.model.delete').subscribe(x => entity = x);
+        this.translate.get('confirm.delete').subscribe(x => message = x);
+        this.translate.get('yes').subscribe(x => yes = x);
+        this.translate.get('no').subscribe(x => no = x);
+
+        if (deviceModel.numberOfDevices == 0) {
+
+            this.messageService
+                .confirm(entity, message, yes, no)
+                .subscribe(confirmed => {
+                    if (confirmed) {
+                        this.deviceService.deleteDeviceModel(String(deviceModel.id)).subscribe(resp => {
+                            this.getDevicesModels()
+                            this.translate.get('success.device.model.delete').subscribe(x => {
+                                this.messageService.success(x)
+                            })
+                        });
+                    }
+                });
+        } else {
+            this.translate.get("device.model.used").subscribe(x=>{
+                this.messageService.error(x)
+            });
+        }
     }
 }
